backend/requirements.txt
Django==4.2.5
djangorestframework==3.14.0
social-auth-core==4.4.2
requests==2.31.0
